新規 - 型は使い回す予定なので、型フォルダにapiを作りそこに格納しておく
