+++ src/types/api/index.ts
@@ -1,0 +1,6 @@
+export type AxiosResut = {
+    status: number;
+    errorCode: string;
+    error: boolean;
+    payload: object;
+};
